refactor: move test_modes macro for reuse
