+++ bytecode/src/function/instructions/add.rs
@@ -133,114 +133,8 @@ mod tests {
     type P = Process;
 
     mod modes {
-        use super::*;
-
-        macro_rules! binary_instruction_test {
-            ($test_name: ident, $instruction: ident, $a: expr, $b: expr, $c: expr, $assert_equal: expr) => {
-                #[test]
-                fn $test_name() {
-                    let a = Value::<P>::from_str($a);
-                    let b = Value::<P>::from_str($b);
-                    let expected = Value::<P>::from_str($c);
-
-                    let registers = Registers::<P>::default();
-                    registers.define(&Register::from_str("r0"));
-                    registers.define(&Register::from_str("r1"));
-                    registers.define(&Register::from_str("r2"));
-                    registers.assign(&Register::from_str("r0"), a);
-                    registers.assign(&Register::from_str("r1"), b);
-
-                    $instruction::from_str("r0 r1 into r2").evaluate(&registers);
-                    let candidate = registers.load(&Register::from_str("r2"));
-                    if $assert_equal {
-                        assert_eq!(expected, candidate);
-                    } else {
-                        // The equality check should fail due to mismatched Modes.
-                        assert_ne!(expected, candidate);
-                    }
-                }
-            };
-        }
-
-        macro_rules! test_modes {
-            ($type: ident, $instruction: ident, $a: expr, $b: expr, $expected: expr) => {
-                mod $type {
-                    use super::*;
-
-                    binary_instruction_test!(
-                        test_public_and_public_is_not_public,
-                        $instruction,
-                        concat!($a, ".public"),
-                        concat!($b, ".public"),
-                        concat!($expected, ".public"),
-                        false
-                    );
-
-                    binary_instruction_test!(
-                        test_public_and_public_is_private,
-                        $instruction,
-                        concat!($a, ".public"),
-                        concat!($b, ".public"),
-                        concat!($expected, ".private"),
-                        true
-                    );
-
-                    binary_instruction_test!(
-                        test_public_and_private_is_not_public,
-                        $instruction,
-                        concat!($a, ".public"),
-                        concat!($b, ".private"),
-                        concat!($expected, ".public"),
-                        false
-                    );
-
-                    binary_instruction_test!(
-                        test_public_and_private_is_private,
-                        $instruction,
-                        concat!($a, ".public"),
-                        concat!($b, ".private"),
-                        concat!($expected, ".private"),
-                        true
-                    );
-
-                    binary_instruction_test!(
-                        test_private_and_public_is_not_public,
-                        $instruction,
-                        concat!($a, ".private"),
-                        concat!($b, ".public"),
-                        concat!($expected, ".public"),
-                        false
-                    );
-
-                    binary_instruction_test!(
-                        test_private_and_public_is_private,
-                        $instruction,
-                        concat!($a, ".private"),
-                        concat!($b, ".public"),
-                        concat!($expected, ".private"),
-                        true
-                    );
-
-                    binary_instruction_test!(
-                        test_private_and_private_is_not_public,
-                        $instruction,
-                        concat!($a, ".private"),
-                        concat!($b, ".private"),
-                        concat!($expected, ".public"),
-                        false
-                    );
-
-                    binary_instruction_test!(
-                        test_private_and_private_is_private,
-                        $instruction,
-                        concat!($a, ".private"),
-                        concat!($b, ".private"),
-                        concat!($expected, ".private"),
-                        true
-                    );
-                }
-            };
-        }
+        use super::Add;
+        use crate::test_modes;
 
         test_modes!(field, Add, "1field", "2field", "3field");
         test_modes!(group, Add, "2group", "0group", "2group");
@@ -257,7 +151,7 @@ mod tests {
         test_modes!(scalar, Add, "1scalar", "2scalar", "3scalar");
     }
 
-    fn check_binary_instruction_test(first: Value<P>, second: Value<P>, expected: Value<P>) {
+    fn check_add_test(first: Value<P>, second: Value<P>, expected: Value<P>) {
         let registers = Registers::<P>::default();
         registers.define(&Register::from_str("r0"));
         registers.define(&Register::from_str("r1"));
@@ -275,14 +169,14 @@ mod tests {
         let one = Value::<P>::from_str("1field.public");
         let two = Value::<P>::from_str("2field.private");
         let three = Value::<P>::from_str("3field.private");
-        check_binary_instruction_test(one, two, three);
+        check_add_test(one, two, three);
     }
 
     #[test]
     fn test_add_group() {
         let two = Value::<P>::from_str("2group.private");
         let zero = Value::<P>::from_str("0group.private");
-        check_binary_instruction_test(two.clone(), zero, two);
+        check_add_test(two.clone(), zero, two);
     }
 
     #[test]
@@ -290,7 +184,7 @@ mod tests {
         let one = Value::<P>::from_str("1scalar.public");
         let two = Value::<P>::from_str("2scalar.private");
         let three = Value::<P>::from_str("3scalar.private");
-        check_binary_instruction_test(one, two, three);
+        check_add_test(one, two, three);
     }
 
     #[test]
@@ -298,7 +192,7 @@ mod tests {
         let negative_one = Value::<P>::from_str("-1i8.public");
         let two = Value::<P>::from_str("2i8.private");
         let one = Value::<P>::from_str("1i8.private");
-        check_binary_instruction_test(negative_one, two, one);
+        check_add_test(negative_one, two, one);
     }
 
     #[test]
@@ -307,7 +201,7 @@ mod tests {
         let max = Value::<P>::from_str(&format!("{}i8.constant", i8::MAX));
         let one = Value::<P>::from_str("1i8.constant");
         let unreachable = Value::<P>::from_str("\"Unreachable\".constant");
-        check_binary_instruction_test(max, one, unreachable);
+        check_add_test(max, one, unreachable);
     }
 
     #[test]
@@ -315,7 +209,7 @@ mod tests {
         let negative_one = Value::<P>::from_str("-1i16.public");
         let two = Value::<P>::from_str("2i16.private");
         let one = Value::<P>::from_str("1i16.private");
-        check_binary_instruction_test(negative_one, two, one);
+        check_add_test(negative_one, two, one);
     }
 
     #[test]
@@ -324,7 +218,7 @@ mod tests {
         let max = Value::<P>::from_str(&format!("{}i16.constant", i16::MAX));
         let one = Value::<P>::from_str("1i16.constant");
         let unreachable = Value::<P>::from_str("\"Unreachable\".constant");
-        check_binary_instruction_test(max, one, unreachable);
+        check_add_test(max, one, unreachable);
     }
 
     #[test]
@@ -332,7 +226,7 @@ mod tests {
         let negative_one = Value::<P>::from_str("-1i32.public");
         let two = Value::<P>::from_str("2i32.private");
         let one = Value::<P>::from_str("1i32.private");
-        check_binary_instruction_test(negative_one, two, one);
+        check_add_test(negative_one, two, one);
     }
 
     #[test]
@@ -341,7 +235,7 @@ mod tests {
         let max = Value::<P>::from_str(&format!("{}i32.constant", i32::MAX));
         let one = Value::<P>::from_str("1i32.constant");
         let unreachable = Value::<P>::from_str("\"Unreachable\".constant");
-        check_binary_instruction_test(max, one, unreachable);
+        check_add_test(max, one, unreachable);
     }
 
     #[test]
@@ -349,7 +243,7 @@ mod tests {
         let negative_one = Value::<P>::from_str("-1i64.public");
         let two = Value::<P>::from_str("2i64.private");
         let one = Value::<P>::from_str("1i64.private");
-        check_binary_instruction_test(negative_one, two, one);
+        check_add_test(negative_one, two, one);
     }
 
     #[test]
@@ -358,7 +252,7 @@ mod tests {
         let max = Value::<P>::from_str(&format!("{}i64.constant", i64::MAX));
         let one = Value::<P>::from_str("1i64.constant");
         let unreachable = Value::<P>::from_str("\"Unreachable\".constant");
-        check_binary_instruction_test(max, one, unreachable);
+        check_add_test(max, one, unreachable);
     }
 
     #[test]
@@ -366,7 +260,7 @@ mod tests {
         let negative_one = Value::<P>::from_str("-1i128.public");
         let two = Value::<P>::from_str("2i128.private");
         let one = Value::<P>::from_str("1i128.private");
-        check_binary_instruction_test(negative_one, two, one);
+        check_add_test(negative_one, two, one);
     }
 
     #[test]
@@ -375,7 +269,7 @@ mod tests {
         let max = Value::<P>::from_str(&format!("{}i128.constant", i128::MAX));
         let one = Value::<P>::from_str("1i128.constant");
         let unreachable = Value::<P>::from_str("\"Unreachable\".constant");
-        check_binary_instruction_test(max, one, unreachable);
+        check_add_test(max, one, unreachable);
     }
 
     #[test]
@@ -383,7 +277,7 @@ mod tests {
         let one = Value::<P>::from_str("1u8.public");
         let two = Value::<P>::from_str("2u8.private");
         let three = Value::<P>::from_str("3u8.private");
-        check_binary_instruction_test(one, two, three);
+        check_add_test(one, two, three);
     }
 
     #[test]
@@ -392,7 +286,7 @@ mod tests {
         let max = Value::<P>::from_str(&format!("{}u8.constant", u8::MAX));
         let one = Value::<P>::from_str("1u8.constant");
         let unreachable = Value::<P>::from_str("\"Unreachable\".constant");
-        check_binary_instruction_test(max, one, unreachable);
+        check_add_test(max, one, unreachable);
     }
 
     #[test]
@@ -400,7 +294,7 @@ mod tests {
         let one = Value::<P>::from_str("1u16.public");
         let two = Value::<P>::from_str("2u16.private");
         let three = Value::<P>::from_str("3u16.private");
-        check_binary_instruction_test(one, two, three);
+        check_add_test(one, two, three);
     }
 
     #[test]
@@ -409,7 +303,7 @@ mod tests {
         let max = Value::<P>::from_str(&format!("{}u16.constant", u16::MAX));
         let one = Value::<P>::from_str("1u16.constant");
         let unreachable = Value::<P>::from_str("\"Unreachable\".constant");
-        check_binary_instruction_test(max, one, unreachable);
+        check_add_test(max, one, unreachable);
     }
 
     #[test]
@@ -417,7 +311,7 @@ mod tests {
         let one = Value::<P>::from_str("1u32.public");
         let two = Value::<P>::from_str("2u32.private");
         let three = Value::<P>::from_str("3u32.private");
-        check_binary_instruction_test(one, two, three);
+        check_add_test(one, two, three);
     }
 
     #[test]
@@ -426,7 +320,7 @@ mod tests {
         let max = Value::<P>::from_str(&format!("{}u32.constant", u32::MAX));
         let one = Value::<P>::from_str("1u32.constant");
         let unreachable = Value::<P>::from_str("\"Unreachable\".constant");
-        check_binary_instruction_test(max, one, unreachable);
+        check_add_test(max, one, unreachable);
     }
 
     #[test]
@@ -434,7 +328,7 @@ mod tests {
         let one = Value::<P>::from_str("1u64.public");
         let two = Value::<P>::from_str("2u64.private");
         let three = Value::<P>::from_str("3u64.private");
-        check_binary_instruction_test(one, two, three);
+        check_add_test(one, two, three);
     }
 
     #[test]
@@ -443,7 +337,7 @@ mod tests {
         let max = Value::<P>::from_str(&format!("{}u64.constant", u64::MAX));
         let one = Value::<P>::from_str("1u64.constant");
         let unreachable = Value::<P>::from_str("\"Unreachable\".constant");
-        check_binary_instruction_test(max, one, unreachable);
+        check_add_test(max, one, unreachable);
     }
 
     #[test]
@@ -451,7 +345,7 @@ mod tests {
         let one = Value::<P>::from_str("1u128.public");
         let two = Value::<P>::from_str("2u128.private");
         let three = Value::<P>::from_str("3u128.private");
-        check_binary_instruction_test(one, two, three);
+        check_add_test(one, two, three);
     }
 
     #[test]
@@ -460,7 +354,7 @@ mod tests {
         let max = Value::<P>::from_str(&format!("{}u128.constant", u128::MAX));
         let one = Value::<P>::from_str("1u128.constant");
         let unreachable = Value::<P>::from_str("\"Unreachable\".constant");
-        check_binary_instruction_test(max, one, unreachable);
+        check_add_test(max, one, unreachable);
     }
 
     #[test]
@@ -473,7 +367,7 @@ mod tests {
         let second = Value::<P>::from_str("4scalar.public");
         let unreachable = Value::<P>::from_str("\"Unreachable\".constant");
 
-        check_binary_instruction_test(composite, second, unreachable);
+        check_add_test(composite, second, unreachable);
     }
 
     #[test]
@@ -483,7 +377,7 @@ mod tests {
         let second = Value::<P>::from_str("4scalar.public");
         let unreachable = Value::<P>::from_str("\"Unreachable\".constant");
 
-        check_binary_instruction_test(invalid_add_literal, second, unreachable);
+        check_add_test(invalid_add_literal, second, unreachable);
     }
 
     #[test]
@@ -494,7 +388,7 @@ mod tests {
         let second = Value::<P>::from_str("4scalar.public");
         let unreachable = Value::<P>::from_str("\"Unreachable\".constant");
 
-        check_binary_instruction_test(invalid_add_literal, second, unreachable);
+        check_add_test(invalid_add_literal, second, unreachable);
     }
 
     #[test]
@@ -504,6 +398,6 @@ mod tests {
         let second = Value::<P>::from_str("4scalar.public");
         let unreachable = Value::<P>::from_str("\"Unreachable\".constant");
 
-        check_binary_instruction_test(invalid_add_literal, second, unreachable);
+        check_add_test(invalid_add_literal, second, unreachable);
     }
 }
